+++ src/mol-gl/shader/chunks/apply-interior-color.glsl.ts
@@ -6,7 +6,7 @@ if (interior) {
         gl_FragColor.rgb *= 1.0 - uInteriorDarkening;
     }
 
-    #ifdef dOpaqueBackfaces
+    #ifdef dTransparentBackfaces_opaque
         gl_FragColor.a = 1.0;
     #endif
 }