+++ src/mol-gl/shader/chunks/check-picking-alpha.glsl.ts
@@ -4,7 +4,7 @@ float fogFactor = smoothstep(uFogNear, uFogFar, abs(viewZ));
 float alpha = (1.0 - fogFactor) * uAlpha;
 // if not opaque enough ignore so the element below can be picked
 if (uAlpha < uPickingAlphaThreshold || alpha < 0.1) {
-    #ifdef dOpaqueBackfaces
+    #ifdef dTransparentBackfaces_opaque
         if (!interior) discard;
     #else
         discard;